pyproject.toml
[tool.poetry]
name = "goit-web-hw11"
version = "0.1.0"
description = ""
authors = ["Andriy414414 <133632261+Andriy414414@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
psycopg2 = "^2.9.9"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.5.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-jwt-auth = "^0.5.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.1.2"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.0"
fastapi-limiter = "^0.1.5"
gravatar = "^0.1"
libgravatar = "^1.0.4"
cloudinary = "^1.37.0"
sphinx = "^7.2.6"
pytest = "^7.4.4"
asyncio = "^3.4.3"
httpx = "^0.26.0"
asqlite3 = "^0.0.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.19.0"
pytest = "^7.4.4"
asyncio = "^3.4.3"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["tests"]
pythonpath = "."
filterwarnings = ["ignore::DeprecationWarning"]